Refactor packages
Updated testing script for new folder structure
