+++ src/main/java/duke/entities/Todo.java
@@ -1,4 +1,4 @@
-package entities;
+package duke.entities;
 
 public class Todo extends Task {
