+++ src/main/java/duke/handlers/EventHandler.java
@@ -1,9 +1,9 @@
-package handlers;
+package duke.handlers;
 
-import entities.Event;
-import entities.Task;
-import exceptions.DukeException;
-import service.Service;
+import duke.entities.Event;
+import duke.entities.Task;
+import duke.exceptions.DukeException;
+import duke.service.Service;
 import java.time.format.DateTimeParseException;
 
 public class EventHandler implements IHandler{
