+++ src/main/java/duke/handlers/IHandler.java
@@ -1,0 +1,8 @@
+package duke.handlers;
+
+import duke.exceptions.DukeException;
+import duke.service.Service;
+
+public interface IHandler {
+    void handle(Service list) throws DukeException;
+}
