--- src/main/java/handlers/IHandler.java
@@ -1,8 +1,0 @@
-package handlers;
-
-import exceptions.DukeException;
-import service.Service;
-
-public interface IHandler {
-    void handle(Service list) throws DukeException;
-}
